Refactor and optimize map view page
