--- src/app/pages/MapViewPage.tsx
@@ -1,344 +1,0 @@
-import {
-  FormControlLabel,
-  Stack,
-  styled,
-  Switch,
-  Tooltip,
-  Typography,
-  useTheme,
-} from "@mui/material";
-import { Fragment, useMemo, useState } from "react";
-import { useHistory } from "react-router";
-import { Directions, PestControlRodent, Place } from "@mui/icons-material";
-import { groupBy } from "lodash";
-import Xarrow, { Xwrapper } from "react-xarrows";
-import { Header } from "../layout/Header";
-import { trpc } from "../state/client";
-import { router } from "../router";
-import { useRouteParams } from "../../lib/hooks/useRouteParams";
-import { MapCoordinate, MapViewport } from "../components/MapViewport";
-import { TabSwitch } from "../components/TabSwitch";
-import { WarpGrid } from "../grids/WarpGrid";
-import { MonsterSpawnGrid } from "../grids/MonsterSpawnGrid";
-import { DataGridQueryFn } from "../components/DataGrid";
-import {
-  MapInfo,
-  Warp,
-  WarpFilter,
-  WarpId,
-} from "../../api/services/map/types";
-import { MapPin } from "../components/MapPin";
-import { Link } from "../components/Link";
-import {
-  MonsterSpawn,
-  MonsterSpawnFilter,
-  MonsterSpawnId,
-} from "../../api/services/monster/types";
-import { defined } from "../../lib/std/defined";
-import { createSwarms } from "../../lib/createSwarms";
-import { center, distance, intersect, Point } from "../../lib/geometry";
-import { LoadingPage } from "./LoadingPage";
-
-export default function MapViewPage() {
-  const history = useHistory();
-  const [showWarpPins, setShowWarpPins] = useState(true);
-  const [showMonsterPins, setShowMonsterPins] = useState(true);
-  const [highlightSpawnId, setHighlightSpawnId] = useState<MonsterSpawnId>();
-  const [highlightWarpId, setHighlightWarpId] = useState<WarpId>();
-  const { id, x, y, tab = "warps" } = useRouteParams(router.map().view);
-  const {
-    data: map,
-    isFetching,
-    isLoading,
-    error,
-  } = trpc.map.read.useQuery(id);
-  const routePoint = definedPoint({ x, y });
-
-  const { data: { entities: warps = [] } = {} } = (
-    trpc.map.searchWarps.useQuery as unknown as DataGridQueryFn<
-      Warp,
-      WarpFilter
-    >
-  )({ filter: { fromMap: { value: id, matcher: "equals" } }, limit: 50 });
-
-  const { data: { entities: spawns = [] } = {} } = (
-    trpc.monster.searchSpawns.useQuery as unknown as DataGridQueryFn<
-      MonsterSpawn,
-      MonsterSpawnFilter
-    >
-  )({
-    filter: {
-      map: { value: id, matcher: "equals" },
-      x: { value: 0, matcher: ">" },
-      y: { value: 0, matcher: ">" },
-    },
-    limit: 50,
-  });
-
-  if (isLoading || isFetching) {
-    return <LoadingPage />;
-  }
-  if (!map || error) {
-    return <Header>Map not found</Header>;
-  }
-
-  return (
-    <>
-      <Header back={router.map}>{map.displayName}</Header>
-      <Stack spacing={2} direction="row" sx={{ flex: 1 }}>
-        <Stack direction="column" sx={{ flex: 2 }}>
-          <Stack direction="row" sx={{ height: 48 }}>
-            <FormControlLabel
-              control={
-                <Switch
-                  checked={showWarpPins}
-                  onChange={(e) => setShowWarpPins(e.target.checked)}
-                />
-              }
-              label="Show Warps"
-            />
-            <FormControlLabel
-              control={
-                <Switch
-                  checked={showMonsterPins}
-                  onChange={(e) => setShowMonsterPins(e.target.checked)}
-                />
-              }
-              label="Show Monsters"
-            />
-          </Stack>
-          <MapView
-            map={map}
-            tab={tab}
-            warps={warps}
-            spawns={spawns}
-            routePoint={routePoint}
-            highlightWarpId={highlightWarpId}
-            highlightSpawnId={highlightSpawnId}
-            setHighlightWarpId={setHighlightWarpId}
-            showWarpPins={showWarpPins}
-            showMonsterPins={showMonsterPins}
-          />
-        </Stack>
-        <Stack direction="column" sx={{ flex: 3 }}>
-          <TabSwitch
-            activeTabId={tab}
-            onChange={(e, newTab) =>
-              history.replace(router.map().view({ id, tab: newTab, x, y }).$)
-            }
-            tabs={[
-              {
-                id: "warps",
-                label: "Warps",
-                content: (
-                  <WarpGrid
-                    filter={{ fromMap: { value: id, matcher: "equals" } }}
-                    onHoveredEntityChange={(entity) =>
-                      setHighlightWarpId(entity?.npcEntityId)
-                    }
-                  />
-                ),
-              },
-              {
-                id: "monsters",
-                label: "Monsters",
-                content: (
-                  <MonsterSpawnGrid
-                    filter={{ map: { value: id, matcher: "equals" } }}
-                    gridProps={{ columnVisibilityModel: { map: false } }}
-                    onHoveredEntityChange={(entity) =>
-                      setHighlightSpawnId(entity?.npcEntityId)
-                    }
-                  />
-                ),
-              },
-            ]}
-          />
-        </Stack>
-      </Stack>
-    </>
-  );
-}
-
-function MapView({
-  map,
-  tab,
-  warps,
-  spawns,
-  highlightWarpId,
-  highlightSpawnId,
-  setHighlightWarpId,
-  routePoint,
-  showWarpPins,
-  showMonsterPins,
-}: {
-  map: MapInfo;
-  tab?: string;
-  warps: Warp[];
-  spawns: MonsterSpawn[];
-  highlightWarpId?: WarpId;
-  setHighlightWarpId?: (id?: WarpId) => void;
-  highlightSpawnId?: MonsterSpawnId;
-  routePoint?: Point;
-  showWarpPins?: boolean;
-  showMonsterPins?: boolean;
-}) {
-  const theme = useTheme();
-
-  const spawnSwarms = useMemo(() => createMonsterSwarms(spawns), [spawns]);
-  const highlightedWarp = warps.find(
-    (warp) => warp.npcEntityId === highlightWarpId
-  );
-  const localWarps = warps.filter(
-    (warp) => warp.toMap === map.id && warp === highlightedWarp
-  );
-  const highlightedSwarm = spawnSwarms.find((swarm) =>
-    swarm.all.some(
-      (spawn) =>
-        spawn.npcEntityId === highlightSpawnId || intersect(spawn, routePoint)
-    )
-  );
-  return (
-    <MapViewport imageUrl={map.imageUrl} bounds={map.bounds}>
-      {showWarpPins && (
-        <Xwrapper>
-          {localWarps.map((warp, index) => (
-            <Fragment key={index}>
-              <Xarrow
-                start={warpXArrowId(warp)}
-                end={pointXArrowId({ x: warp.toX, y: warp.toY })}
-                color={theme.palette.primary.main}
-              />
-              <MapCoordinate
-                id={pointXArrowId({ x: warp.toX, y: warp.toY })}
-                x={warp.toX}
-                y={warp.toY}
-              />
-            </Fragment>
-          ))}
-          {warps.map((warp, index) => {
-            const mouseBindings = {
-              onMouseOver: () => setHighlightWarpId?.(warp.npcEntityId),
-              onMouseOut: () => setHighlightWarpId?.(undefined),
-            };
-            return (
-              <MapPin
-                data-testid="Map pin"
-                key={`warp${index}`}
-                x={warp.fromX}
-                y={warp.fromY}
-                highlight={warp === highlightedWarp}
-                {...mouseBindings}
-                label={
-                  <LinkOnMap
-                    to={router.map().view({
-                      id: warp.toMap,
-                      x: warp.toX,
-                      y: warp.toY,
-                      tab,
-                    })}
-                  >
-                    <MapPinLabel {...mouseBindings} color="white">
-                      {warp.toMap}
-                    </MapPinLabel>
-                  </LinkOnMap>
-                }
-              >
-                <MapPinIcon id={warpXArrowId(warp)} sx={mapPinIconCss} />
-              </MapPin>
-            );
-          })}
-        </Xwrapper>
-      )}
-      {showMonsterPins &&
-        spawnSwarms.map((swarm, index) => (
-          <MapPin
-            key={index}
-            x={swarm.x}
-            y={swarm.y}
-            highlight={swarm === highlightedSwarm}
-            label={
-              <>
-                {swarm.groups.map((group, index) => (
-                  <LinkOnMap
-                    key={index}
-                    to={router.monster().view({ id: group.id })}
-                    sx={{ lineHeight: "1em" }}
-                  >
-                    <MapPinLabel color={monsterColor}>
-                      {group.name} {group.size > 1 ? `x${group.size}` : ""}
-                    </MapPinLabel>
-                  </LinkOnMap>
-                ))}
-              </>
-            }
-          >
-            <PestControlRodent sx={{ ...mapPinIconCss, color: monsterColor }} />
-          </MapPin>
-        ))}
-      {routePoint && (
-        <MapCoordinate x={routePoint.x} y={routePoint.y}>
-          <Tooltip title="The warp you selected leads here">
-            <Place
-              sx={{ ...mapPinIconCss, fill: theme.palette.success.main }}
-            />
-          </Tooltip>
-        </MapCoordinate>
-      )}
-    </MapViewport>
-  );
-}
-
-const monsterColor = "#ff7878";
-
-const mapPinIconCss = {
-  color: "#fff",
-  filter: `drop-shadow( 0 0 1px rgba(0, 0, 0, 1))`,
-};
-
-const LinkOnMap = styled(Link)`
-  text-decoration: none;
-  display: flex;
-`;
-
-const MapPinIcon = Directions;
-
-const MapPinLabel = styled(Typography)`
-  line-height: 1em;
-  font-size: ${(p) => p.theme.typography.caption.fontSize};
-  text-shadow: -1px -1px 0 #000, 1px -1px 0 #000, -1px 1px 0 #000;
-`;
-
-function createMonsterSwarms(spawns: MonsterSpawn[], swarmDistance = 15) {
-  const spawnsWithLocations = defined(
-    spawns.map(({ x, y, ...props }) =>
-      x !== undefined && y !== undefined ? { ...props, x, y } : undefined
-    )
-  );
-  const swarms = createSwarms(
-    spawnsWithLocations,
-    (a, b) => distance(a, b) <= swarmDistance
-  );
-  return swarms.map((swarm) => {
-    return {
-      ...center(swarm),
-      all: swarm,
-      groups: Object.values(groupBy(swarm, (spawn) => spawn.id)).map(
-        (group) => ({
-          name: group[0].name,
-          id: group[0].id,
-          size: group.length,
-        })
-      ),
-    };
-  });
-}
-
-function definedPoint(point?: Partial<Point>) {
-  return point?.x !== undefined && point?.y !== undefined
-    ? { x: point.x, y: point.y }
-    : undefined;
-}
-
-const warpXArrowId = (warp: Warp) => `warp_arrow_${warp.npcEntityId}`;
-const pointXArrowId = (point: Point) => `point_arrow_${point.x}_${point.y}`;
